+++ CHANGELOG.md
@@ -3,15 +3,46 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## v2.2.0dev - [name] - [date]
+## v2.3dev - 2023-08-29
+
+### `Changed`
+
+- [#213](https://github.com/nf-core/epitopeprediction/pull/203) - Rename param `genome_version` to `genome_reference`, add functionality to handle BioMart archive urls
+- [#213](https://github.com/nf-core/epitopeprediction/pull/203) - Update to nf-core template `2.10`
+- [#203](https://github.com/nf-core/epitopeprediction/pull/203) - Update to nf-core template `2.9`
+- [#203](https://github.com/nf-core/epitopeprediction/pull/203) - Update to nf-core template `2.8`
+- [#206](https://github.com/nf-core/epitopeprediction/issues/206) - Update the row checker class.
+
+### `Fixed`
+
+- [#219](https://github.com/nf-core/epitopeprediction/pull/219) - Fix `EXTERNAL_TOOLS_IMPORT`` container registry and bump version
+
+## v2.2.1 - WaldhaeuserOst Hotfix - 2023-03-16
+
+### `Fixed`
+
+- [#196](https://github.com/nf-core/epitopeprediction/pull/196) - Revert versions changes that caused bug with external tools predictions missing.
+
+## v2.2.0 - WaldhaeuserOst - 2023-03-03
 
 ### `Added`
+
+- [#180](https://github.com/nf-core/epitopeprediction/pull/180) - Add support for `VEP` annotated VCF files [#172](https://github.com/nf-core/epitopeprediction/issues/172)
+- [#186](https://github.com/nf-core/epitopeprediction/pull/186) - Log messages from `epaa.py` script to stdout and provide `sys.exit` error messages.
 
 ### `Changed`
 
 - [#177](https://github.com/nf-core/epitopeprediction/pull/177) - Update to nf-core template `2.5.1`
+- [#178](https://github.com/nf-core/epitopeprediction/pull/178) - Update MultiQC to `1.13`
+- [#180](https://github.com/nf-core/epitopeprediction/pull/180) - Update to nf-core template `2.6`
+- [#180](https://github.com/nf-core/epitopeprediction/pull/180) - Improve runtime for VCF-based predictions
+- [#187](https://github.com/nf-core/epitopeprediction/pull/187) - Update to nf-core template `2.7.1`
+- [#189](https://github.com/nf-core/epitopeprediction/pull/189) - Update to nf-core template `2.7.2`
 
 ### `Fixed`
+
+- [#180](https://github.com/nf-core/epitopeprediction/pull/180) - Fix issue with `frameshift` determination
+- [#194](https://github.com/nf-core/epitopeprediction/pull/194) - Fix software versions collection and add script licenses
 
 ## v2.1.0 - Nordring - 2022-08-02